+++ packages/user/src/supertokens/recipes/config/third-party-email-password/thirdPartySignInUpPost.ts
@@ -8,7 +8,7 @@ const thirdPartySignInUpPOST = (
   originalImplementation: APIInterface,
   fastify: FastifyInstance
 ): typeof originalImplementation.thirdPartySignInUpPOST => {
-  const { log } = fastify;
+  const { log, config } = fastify;
 
   return async (input) => {
     if (originalImplementation.thirdPartySignInUpPOST === undefined) {
@@ -21,7 +21,7 @@ const thirdPartySignInUpPOST = (
     if (originalResponse.status === "OK" && originalResponse.createdNewUser) {
       const rolesResponse = await UserRoles.addRoleToUser(
         originalResponse.user.id,
-        "USER"
+        config.user.role || "USER"
       );
 
       if (rolesResponse.status !== "OK") {